requirements.txt
Flask~=1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
Werkzeug==1.0.1
praw~=7.1.4
python-dotenv~=0.15.0
Flask-Cors==3.0.10